# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor Appointment API",
        "description": "API Documentation for Doctor Appointment System",
        "contact": {
            "email": "ujjwal999adhikari@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Doctor Appointment API Server"
        }
    ],
    "paths": {
...
        "/doctor/store": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Store a newly created doctor's information",
                "description": "Store a doctor's information and return the stored data along with a success message",
                "operationId": "96def4a2d4f11c8823f3beac2a1f8e43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doctor_description",
                                    "doctor_phone",
                                    "department_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "doctor_description": {
                                        "type": "string",
                                        "example": "Experienced Cardiologist with 10 years of practice."
                                    },
                                    "doctor_phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor information stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "doctor": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "doctor_description": {
                                                    "type": "string",
                                                    "example": "Experienced Cardiologist with 10 years of practice."
                                                },
                                                "doctor_phone": {
                                                    "type": "string",
                                                    "example": "+1234567890"
                                                },
                                                "department_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-08-27T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-08-27T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Doctor information stored successfully"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "success"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Doctor data already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You already have inserted your data"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "doctor_description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The doctor_description field is required."
                                                    }
                                                },
                                                "doctor_phone": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The doctor_phone field is required."
                                                    }
                                                },
                                                "department_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The department_id field is required."
                                                    }
                                                },
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The user_id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: `Bearer {token}`",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}